src/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
iniconfig==2.0.0
invoke==2.2.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0
urllib3==2.3.0
